ty type checking: fix warnings in pro and hs, add hs dev environment
Pro:
- Fix 5 invalid-return-type warnings from OpenCV imprecise stubs
  - processing/color.py: .astype(np.uint8) on cvtColor and LUT returns
  - processing/resize.py: np.asarray(..., dtype=np.uint8) at final return
  - capture/opencv.py: np.asarray(frame, dtype=np.uint8) on cap.read() return
- Bump default baud_rate 2M -> 4M in config.py, default.yaml, test_config.py
- Remove stale comment from pi.yaml
- Update README: test count 136->159, expand ty section with config notes

HS:
- Add hs/pyproject.toml for developer tooling (ty, ruff) - students unaffected
- Fix 3 invalid-argument-type errors: send_frame, run_snapshot,
  run_avatar_capture all accept Optional[serial.Serial] (already handled
  None at runtime, signatures now match)
- Add Developer/Maintainer section to README with ty and ruff usage

Gitignore:
- Add avatar_*/ pattern to ignore avatar capture artifact directories

+++ pro/src/ledportal_pro/processing/color.py
@@ -47,7 +47,7 @@ def apply_grayscale(frame: NDArray[np.uint8]) -> NDArray[np.uint8]:
         BGR image where R=G=B (grayscale appearance).
     """
     gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-    return cv2.cvtColor(gray, cv2.COLOR_GRAY2BGR)
+    return cv2.cvtColor(gray, cv2.COLOR_GRAY2BGR).astype(np.uint8)
 
 
 def apply_brightness_limit(frame: NDArray[np.uint8], max_brightness: int = 255) -> NDArray[np.uint8]:
@@ -88,7 +88,7 @@ def apply_gamma_correction(frame: NDArray[np.uint8], gamma: float = 2.2) -> NDAr
     inv_gamma = 1.0 / gamma
     table = np.array([((i / 255.0) ** inv_gamma) * 255 for i in range(256)]).astype(np.uint8)
 
-    return cv2.LUT(frame, table)
+    return cv2.LUT(frame, table).astype(np.uint8)
 
 
 def rgb565_to_bgr(data: bytes, width: int, height: int) -> NDArray[np.uint8]:
